+++ src/Makefile
@@ -1649,6 +1649,7 @@ BASIC_SRC = \
 	main.c \
 	map.c \
 	mark.c \
+	match.c \
 	mbyte.c \
 	memfile.c \
 	memline.c \
@@ -1797,6 +1798,7 @@ OBJ_COMMON = \
 	objects/list.o \
 	objects/map.o \
 	objects/mark.o \
+	objects/match.o \
 	objects/mbyte.o \
 	objects/memline.o \
 	objects/menu.o \
@@ -1971,6 +1973,7 @@ PRO_AUTO = \
 	main.pro \
 	map.pro \
 	mark.pro \
+	match.pro \
 	mbyte.pro \
 	memfile.pro \
 	memline.pro \
@@ -3379,6 +3382,9 @@ objects/map.o: map.c
 objects/mark.o: mark.c
 	$(CCC) -o $@ mark.c
 
+objects/match.o: match.c
+	$(CCC) -o $@ match.c
+
 objects/memfile.o: memfile.c
 	$(CCC) -o $@ memfile.c
 
@@ -3963,6 +3969,10 @@ objects/map.o: map.c vim.h protodef.h auto/config.h feature.h os_unix.h \
  proto.h globals.h
 objects/mark.o: mark.c vim.h protodef.h auto/config.h feature.h os_unix.h \
  auto/osdef.h ascii.h keymap.h term.h macros.h option.h beval.h \
+ proto/gui_beval.pro structs.h regexp.h gui.h alloc.h ex_cmds.h spell.h \
+ proto.h globals.h
+objects/match.o: match.c vim.h protodef.h auto/config.h feature.h \
+ os_unix.h auto/osdef.h ascii.h keymap.h term.h macros.h option.h beval.h \
  proto/gui_beval.pro structs.h regexp.h gui.h alloc.h ex_cmds.h spell.h \
  proto.h globals.h
 objects/mbyte.o: mbyte.c vim.h protodef.h auto/config.h feature.h os_unix.h \
